Move the implementation of `syntactic_environment` to `.cpp` file
